refactor(compass-aggregations): address PR review feedback on COMPASS-10832
- AnalyzeAndRefineResultsButton now calls useDarkMode() itself instead of
  requiring callers to compute and pass it as a prop.
- Extract the duplicated pipeline-string-building logic from
  stage-preview/index.tsx and focus-mode-stage-preview.tsx's connect()
  calls into a shared buildPipelineStringUpToStage() helper.
